Fixes Ody phyinit SBE compile by updating constexpr to uint32_t
This fixes a misaligned access issue and was missed in a backport

Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149932
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/149949
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>

+++ src/import/chips/ocmb/odyssey/procedures/hwp/memory/lib/phy/ody_ddrphy_phyinit_config.C
@@ -3864,7 +3864,7 @@ fapi2::ReturnCode setup_phy_advanced_struct(const fapi2::Target<fapi2::TARGET_TY
 
     // ATxImpedance (set to DIMM0, rank0 value)
     {
-        constexpr uint16_t ATxIMPCMD_Arr[mss::ody::sizes::SPD_ATXCMD_DECODE_MAX]
+        constexpr uint32_t ATxIMPCMD_Arr[mss::ody::sizes::SPD_ATXCMD_DECODE_MAX]
         __attribute__ ((aligned (8))) =
         {
             // These values are pulled from the DDR5 DDIMM SPD doc rev.1.21 for EFD Bytes 32
